Issue #1343874 by jm1: node_reference support: too many image nodes displayed
This replaces a long overdue workaround which loaded the images for a node_reference gallery directly from the database instead of using the Field API.
It's now possible to select multiple image/media fields to show when formatting a node_reference field.
At the same time we got rid of the 'galleria_' prefix for the field formatter settings which requires an update.php run to adopt the old settings.